Simplify: defer how renderable items are given the registry.
Template content item structs no longer have their own reference to the
handlebars registry. Instead it is provided at render time.

RenderArgs now contains both the RenderData (template parameters) and a
reference to the ContentEngine.

Rc<_> wrapping of the handlebars registry could likely be removed as
well, but that is not done in this commit.