Implement `get` handlebars helper.
This embeds other content (both static content and other templates)
into the current template. It's somewhat like a generalization of the
handlebars "partial" concept to also work for static content.

This includes a nontrivial change to ContentEngine's constructor. It
now returns an Arc<RwLock<ContentEngine>> upon success instead of a
plain ContentEngine. This is because the `get` helper needs to capture
a reference to the engine, and that requires Sync + Send. I have a
feeling there's some way to eliminate the lock (since it only needs to
be mutated once during initialization), but that will be future work.