Fix media type handling of `get` helper.
Previously it used the target media type requested by the end user,
when it really should have been using the source media type of the
including template. This happens to work given the limited set of
functionality that's been implemented so far, but this won't be true
for long.

Some cleanup is still needed, e.g. `RenderData.target_media_type`
should be removed.