Pass a list of acceptable media ranges to render calls.
This replaces the single target media type. Currently only the first
one is looked at in all Render impls, but the idea is to work down the
list and see if any of the media ranges can work out.

The list is expected to be ordered from most-preferred to
least-preferred.