Change handling of extensions on URL paths.
As a user, turns out it's weird to see http://site.com/lolcat.pdf work
when it's clearly a gif.

So now if there is a URL extension the accept header value is
completely ignored. That is, the extension, if present, specifies the
ONLY acceptable type instead of the MOST acceptable type.