Filter out hidden paths in content directories earlier on.
This can speed up initialization since we don't have to traverse into
hidden directories, and helps cut down on the number of file
descriptors that are open at the same time.

The only time I've noticed issues is when there's .git metadata in the
content directory, but that'll probably be common.