New feature: unrouteable content files.
With this change, files whose names start with underscore (e.g.
`_foo.html`) are no longer accessible from user requests. They're
still part of the registry and can be included in templates using the
`get` helper or partial syntax.

This is useful for partials like headers, footers, etc that shouldn't
be served out of context. Prior to this change you could `curl
http://website.com/header` which was pretty weird.