Remove a redundant category from package metadata.
After seeing what it looks like on crates.io, I've decided that since
Operator has the "web-programming::http-server" category, also having
"network-programming" is not useful and may even be confusing. Out of
the available options I think "web-programming::http-server" plus
"command-line-utilities" is the maximally-informative set of categories
for this package.