WIP Upgrade dependencies.
I'm still waiting on actix-web 4.x and handlebars-rust 4.x and am not
planning to merge this until after they are released. I've played
around with the betas of both libraries and there will definitely be
some code changes required to accommodate them, but am not going to
invest any significant effort in the upgrades until there are release
candidates.