Handle query parameters, including them in render data.
Query strings on request URIs are parsed and provided to templates
and executables as render data at `request.query-parameters`. This
value will typically be a map of key/value pairs.

This commit also backwards-incompatibly relocates the request route's
placement in render data from the top-level `request-route` property
to `request.route`.